requirements.txt
google-auth-oauthlib>=0.4.6
google-auth>=2.3.3
google-api-python-client>=2.31.0
pandas>=1.3.4
pytz>=2021.3
google-api-python-client-stubs>=1.12.0  # For Forms API type hints
Flask>=2.0.0 